Added `useGetClientOperatorsActions` hook
